+++ tests/test_inflector_ca.py
@@ -21,7 +21,7 @@ def test_all_verbs_have_templates():
         if verb.template not in template_names:
             missing_templates.add(verb.template)
             missing_verbs.append(verb)
-    # assert len(missing_templates) == 0
+    assert len(missing_templates) == 0
     # assert len(missing_verbs) == 0
 
 
@@ -2035,6 +2035,21 @@ def test_find_verb_by_infinitive():
             True,
             "m",
             ["evacua", "evacuï", "evacuem", "evacueu", "evacuïn"],
+        ),
+        (
+            "tondre",
+            "indicatiu",
+            "present",
+            True,
+            "f",
+            [
+                "jo tonc",
+                "tu tons",
+                "ella ton",
+                "nosaltres tonem",
+                "vosaltres toneu",
+                "elles tonen",
+            ],
         ),
     ],
 )