+++ verbecc/inflector_pt.py
@@ -108,8 +108,17 @@ class InflectorPt(inflector.Inflector):
     ) -> str:
         return "ter"
 
+    def _get_infinitive_mood_name(self):
+        return "infinitivo"
+
+    def _get_indicative_mood_name(self):
+        return "indicativo"
+
     def _get_subjunctive_mood_name(self) -> str:
         return "subjuntivo"
+
+    def _get_conditional_mood_name(self):
+        return "condicional"
 
     def _get_participle_mood_name(self) -> str:
         return "particípio"