+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - 1.9.5 [24 December 2023]
   - Improved Catalan Support
     - Added more missing templates (2)
-    - TODO: Still missing 24 templates for 85 out of 8616 verbs
+    - TODO: Still missing 23 templates for 84 out of 8616 verbs
     - See `test_inflector_ca.test_all_verbs_have_templates`
 
 - 1.9.4 [23 December 2023]