+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - 1.9.5 [24 December 2023]
   - Improved Catalan Support
     - Added more missing templates (5+)
-    - TODO: Still missing 18 templates for 73 out of 8616 verbs
+    - TODO: Still missing 17 templates for 71 out of 8616 verbs
     - See `test_inflector_ca.test_all_verbs_have_templates`
 
 - 1.9.4 [23 December 2023]