+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - 1.9.5 [24 December 2023]
   - Improved Catalan Support
     - Added more missing templates (6+)
-    - TODO: Still missing 15 templates for 60 out of 8616 verbs
+    - TODO: Still missing 14 templates for 59 out of 8616 verbs
     - See `test_inflector_ca.test_all_verbs_have_templates`
 
 - 1.9.4 [23 December 2023]