+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - 1.9.6 [25 December 2023]
   - Improved Catalan Support
     - Added more missing templates (1)
-    - TODO: Still missing 8 templates for 18 out of 8616 verbs
+    - TODO: Still missing 7 templates for 17 out of 8616 verbs
     - See `test_inflector_ca.test_all_verbs_have_templates`
 
 - 1.9.5 [24 December 2023]