+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # verbecc Changelog
 
-- 1.9.6 [25 December 2023]
+- 1.9.6 [26 December 2023]
   - Improved Catalan Support
-    - Added more missing templates (1)
-    - TODO: Still missing 5 templates for 9 out of 8616 verbs
+    - Added remaining missing templates for all 8616 verbs
     - See `test_inflector_ca.test_all_verbs_have_templates`
 
 - 1.9.5 [24 December 2023]